+++ src/main/kotlin/uk/co/ben_gibson/git/link/url/template/UrlTemplates.kt
@@ -50,15 +50,7 @@ data class UrlTemplates(val fileAtBranch: String, val fileAtCommit : String, val
             )
         }
 
-        fun gitee() = UrlTemplates.gitHub()
-
-        fun azure(): UrlTemplates {
-            return UrlTemplates(
-                "{remote:url}?version=GB{branch}&path=/{file:path}/{file:name}{line-block:start}&line={line:start}&lineEnd={line:end}{line-block:end}&lineStartColumn=1&lineEndColumn=1",
-                "{remote:url}?version=GC{commit}&path=/{file:path}/{file:name}{line-block:start}&line={line:start}&lineEnd={line:end}{line-block:end}&lineStartColumn=1&lineEndColumn=1",
-                "{remote:url}/commit/{commit}"
-            )
-        }
+        fun gitee() = gitHub()
 
         fun gerrit(): UrlTemplates {
             return UrlTemplates(